refactor(containers): move abstract containers to 'common' package
