+++ src/tomodachi_testcontainers/containers/localstack.py
@@ -3,7 +3,7 @@ from typing import Any, Optional
 
 from testcontainers.core.waiting_utils import wait_for_logs
 
-from tomodachi_testcontainers.containers.web import WebContainer
+from tomodachi_testcontainers.containers.common import WebContainer
 from tomodachi_testcontainers.utils import AWSClientConfig
