+++ src/tomodachi_testcontainers/pytest/assertions.py
@@ -1,7 +1,7 @@
 from datetime import datetime, timedelta, timezone
 from typing import Tuple, cast
 
-from tomodachi_testcontainers.containers.common import DockerContainer
+from tomodachi_testcontainers import DockerContainer
 
 UUID4_PATTERN = r"[0-9a-f]{8}-?[0-9a-f]{4}-?4[0-9a-f]{3}-?[89ab][0-9a-f]{3}-?[0-9a-f]{12}"
