+++ tests/containers/common/test_database_container.py
@@ -1,7 +1,7 @@
 import pytest
 from sqlalchemy.exc import OperationalError
 
-from tomodachi_testcontainers.containers.database import DatabaseURL, wait_for_database_healthcheck
+from tomodachi_testcontainers.containers.common.database import DatabaseURL, wait_for_database_healthcheck
 from tomodachi_testcontainers.containers.mysql import MySQLContainer
 from tomodachi_testcontainers.utils import get_available_port
